fix(veeam): fix license float parsing, add discriminant tags registry
- Change license instance/socket counters from int to float64
  (Veeam API returns 85.0 not 85)
- Register veeam discriminant tags (job_name, job_type, repo_name, proxy_name)
  to fix cache key warnings

+++ internal/agent/probes/veeam/types.go
@@ -69,15 +69,16 @@ type licenseInfo struct {
 }
 
 // instanceLicenseSummary holds instance-based license counters
+// Note: Veeam API returns these as floats (e.g. 85.0) not ints
 type instanceLicenseSummary struct {
-	LicensedInstancesNumber int `json:"licensedInstancesNumber"`
-	UsedInstancesNumber     int `json:"usedInstancesNumber"`
+	LicensedInstancesNumber float64 `json:"licensedInstancesNumber"`
+	UsedInstancesNumber     float64 `json:"usedInstancesNumber"`
 }
 
 // socketLicenseSummary holds socket-based license counters
 type socketLicenseSummary struct {
-	LicensedSocketsNumber int `json:"licensedSocketsNumber"`
-	UsedSocketsNumber     int `json:"usedSocketsNumber"`
+	LicensedSocketsNumber float64 `json:"licensedSocketsNumber"`
+	UsedSocketsNumber     float64 `json:"usedSocketsNumber"`
 }
 
 // proxy represents a Veeam backup proxy from /api/v1/backupInfrastructure/proxies
